Update `FindDsl` to work with composite primary keys
This introduces a new expression method, `eq_all` which is not in the
public API. For single values, it is identical to `eq`. For tuples, it
represents all of the tuple elements matching.

I unfortunately could not implement `eq_all` for all types, as the base
impl would always overlap with the tuple impls. Since this is meant to
be used with `FindDsl` for now, which always has a column as the left
hand side, we brute force it.

Additionally, I couldn't easily implement this for arbitrarily large
tuples with a macro. This might get a bit easier when macro types lands
(but the code might still be complex enough to not warrant it). The
overwhelming majority of cases here will be 2 element tuples. I've
manually done it for 4 elements just to be safe.