Re-implement `#[derive(Associations)]` using Macros 1.1
Going forward Macros 1.1 is the intended path of stabilization for
procedural macros, so `diesel_codegen` will need to be rewritten to use
it.

Much of the helper code around this is a direct port of the libsyntax
version of our code, rewritten to use `syn` instead.

We're cloning the attributes, as `quote!` only works for repetition with
`T where for<'a> &'a T: IntoIterator`. I think this is a bug and will
address upstream.