Reimplement `embed_migrations!` using Macros 1.1
There wasn't a ton of code that could be shared between the old crate
and the new one, as the majority of the lines of code are just the setup
of the structs and functions used, which need to go into the respective
quote macro. In order to share that code, it'd have to be moved into
Diesel proper and out of the macro output, which I want to avoid. Since
the syntex crate will be removed once Macros 1.1 is stable, we can live
with the duplication for now.