Release v0.8.0
This release is primarily focused around diesel_codegen, which has been
entirely rewritten to use the new Macros 1.1 framework. This means that
Diesel should continue to work on all future nightlies, and will be much
more compatible with other crates which use codegen such as serde. This
did reqiure some breaking changes to their API, but it should be a
straightforward migration. [the CHANGELOG][changelog] has migration
instructions.

The macros 1.1 rewrite took the majority of our time, so this release is
fairly light on features, but we do have a few great quality of life
additions. See [the CHANGELOG][changelog] for details.

[changelog]: https://github.com/diesel-rs/diesel/blob/v0.8.0/CHANGELOG.md

A huge thank you to all the contributors who worked on this release:

- Cengiz Can
- Christopher Brickley
- David Szotten
- Georg Semmler
- Jimmy Cuadra
- Josh Holmer
- Rasmus Kaj
- Robert Maloney
- Sebastian Blei