Release v0.8.2
This release supports the changes made to the Macros 1.1 system as of
2016-11-07. In order to support this, a breaking change had to be made
to the API of `infer_table_from_schema!`. If you are using this macro,
please see the changelog for more details. If you are not using this
macro, you will be unaffected by this release as long as you are on a
nightly later than 2016-11-07.