Remove the `FindDsl` constraint from `Identifiable`
Turns out we don't rely on it in as many places as we used to, and its
presence is making other refactorings difficult. I think I originally
added it to make the where clause on the impl of `SaveChangesDsl` less
gnarly, but it turns out it's more or less exactly the same anyway.