Separate the types of foreign and primary keys from belongs_to
As part of the attempt to allow composite primary keys in
`Identifiable`, we either need ATCs or reference the lifetime at the
trait impl level either by adding it to the `Identifiable` trait or by
implementing the trait on a reference. However, no matter what I did I
would always end up with a lifetime error in [`grouped_by`][grouped-by]
about the child being borrowed for too long.

I've finally realized that the reason this was occurring was that we
were [insisting that the foreign key be the same type as the primary
key][fk-type]. When we bring an additional lifetime into the picture,
what we were effectively doing is "downcasting" the foreign key to be of
the type of the primary key. This ended up implying that the child had
to live at least as long as the hash map containing the primary keys.
What we're doing now is effectively upcasting the primary key when we do
the comparison, allowing the foreign key to live for a short as it needs
to.

This still does not do the actual lifting of the reference of the
primary key type, that will come following this piece, assuming I don't
run into any more walls.

[grouped-by]: https://github.com/diesel-rs/diesel/blob/be55e4f81471bac57f9cffd42355b64f7ab8d71a/diesel/src/associations/belongs_to.rs#L30-L32
[fk-type]: https://github.com/diesel-rs/diesel/blob/be55e4f81471bac57f9cffd42355b64f7ab8d71a/diesel/src/associations/belongs_to.rs#L11