Add support for `NOT IN`
I will probably eventually re-implement this with an arbitrary `not`
function, but I didn't want to go there quite yet as I'm not sure if
we'll need to do anything with parenthesis for that.

This will add another case we have to deal with empty arrays for, but I
don't think this will affect the implementation other than requiring an
additional test. What I'm thinking we'll do is have the `In` expression
write out `1=0` when it sees an empty array. With this implementation
it'll be `NOT 1=0`, which is exactly what we want. (In empty set = no
rows, not in empty set = all rows)

Fixes #511