Allow custom schema names in the `table!` macro
This does not add support for `infer_schema!` with tables on schemas
other than public, but it does lay the ground work by changing the
macros to support it. I removed the `name` function from the `Table`
trait, as any code that was using that function would be broken on
tables with custom schemas.

Partially fixes #447