Give a more helpful error on no-op updates
Prior to this commit, an update with no changes would generate invalid
SQL which would result in a database error. Even if the resulting error
were handled, the transaction would be aborted at that point, which may
not be desired. With this change, we provide a more useful error message
and don't risk aborting the current transaction.