Fix build on latest nightlies
`use super::*` is now including private members, which is causing a
collision between `diesel::types::Duration` and `std::time::Duration`
(our code is wanting to reference the latter). This is probably a
regression on Rust's end, as it caused code to stop compiling from
stable to nightly. However, it was an intentional change. So even if
this code is changed to compile again in the short term, it will likely
start warning and then be changed in the future anyway. This protects us
against that case for the future.