Generate `foo NOT IN ...` not `NOT foo IN ...`
This means the code is a bit more copy-pasta, but not to a level that
concerns me. This approach felt better (and simpler) to me than unifying
these two AST nodes into something with an operator, since the operator
would need special methods for what to generate when empty.

In addition to generating `foo NOT IN` instead of `NOT foo IN`, we now
also generate `1=1` on an empty array, instead of `NOT 1=0`.

This change has no tests, as the two are functionally equivalent (in
fact, both SQLite and PG will generate identical query plans for both
forms). I prefer to avoid directly testing generated SQL unless it's
absolutely necessary.