impl `ToSql<Nullable<Array<ST>> for &[T]`
We provide a `ToSql` impl for every other type to the nullable version
of its SQL type, but we missed arrays which don't use the same macros as
other types.

We had to brute force this impl previously due to a coherence issue, but
I should revisit this and see if the introduction of `NotNull` has
changed anything. My guess is that we still can't because we have the
`impl<ST, T, DB> ToSql<Nullable<ST>, DB> for Option<T>`, and someone
could in theory provide an impl of `ToSql<SomeType, DB> for
ExternalStruct` where `SomeType: NotNull`, which would mean we need
specialization.