Don't use libpq for identifier escaping
The implementation of `PQescapeIdentifier` actually does very little,
especially if we can assume that the connection's client encoding is
UTF-8, and that the identifier is valid UTF-8. Since I'd like to share
as much code between the current PG adapter, and the new async one, we
can just re-implement this function ourselves.

This was also the only place that required an `Rc` on the
`RawConnection`, which isn't a source of any major overhead but I'm glad
to see it go.

For some reason I thought that replace returned a `Cow<str>`, not
`String`. I may add our own version since this function could avoid
allocating the majority of the time.