Clean up `print-schema` whitespace output
There should be a leading space before the primary key. The intention of
`users(id)` is not to say "this is a function being called with `id`",
it's "the table is called users and also here's an afterthought about it
involving `id`. If we're going to bother prettifying the output, it
should follow the style we document.