Clean up output of `diesel print-schema`
Prior to this commit, the output would include the `mod infer_table {
table! { ... }; pub use self::infer_table::*;` wrapper that we generate.
This is only needed due to how the macros work with Macros 1.1. When
calling the various functions directly, we should generate code that is
closer to what the user would write themselves.

In addition to not including that wrapper code, I've also inserted a
newline between each table! macro, and trimmed all trailing whitespace,
making it nearly exactly what the user would write themselves.