Remove the implicit `'static` bound from `run_migrations`
Rust has the undocumented behavior of implying `+ 'static` any time you
reference a trait object directly. Normally this isn't a problem since
the only time you'd usually do that is as `Box<Trait>`, where you can
just write `Box<Trait + 'a>` instead. We can't do that here, however,
sicne `Deref<Target=Trait + 'a>` is not valid syntax. I'm not sure if
there's a workaround to keep the old code that I'm missing, or if this
is legitimately a problem in the language.

Either way it doesn't matter, we can just implement the trait on box and
ref and everything is fine.

I did not include a changelog entry, as all previous callers of this
function should continue to work, and this is a function that is
documented as "you probably should be using something else"

Fixes #593.