Add initial structure for MySQL support
This represents the bare minimum required to define all the basic
structures MySQL will be needing, and have the tests compile. I've opted
not to include this as a travis allowed failure or in `bin/test` until
at least the core crate is mostly complete.

A few things that stood out to me while going through all of this:

It's called `Mysql` and not `MySql` because the module and feature are
called `mysql` not `my_sql`. The struct would never be called `MySQL` in
ideomatic Rust.

I'm not sure I generated bindings correctly since there's a lot of
different header files, so I haven't published yet. I was rather
surprised nobody has written a wrapper around their C API yet. Just the
pure Rust client (which mirrors rust-postgres's API so will be
impossible to use for all the same reasons)

The C API has global initialization. I believe that wrapping it in a
`std::sync::Once` and calling that from `MysqlConnection::establish`.

Lol if your enum variant for a 32 bit integer is called `Long`

The bind type variants for input vs output are asymmentrical. We are
mirroring the input, as we would only ever care about the output for
debug purposes. There are a lot of types missing from the input list.
Notably, decimal is missing. Unsure how we transmit in that case
(granted, it's not like we have any Rust type to work with anyway)

SqliteQueryBuilder, DebugQueryBuilder, and MysqlQueryBuilder are all
currently identical. I'm not sure that I ever expect them to diverge,
and will consider unifying them.