Don't pass pointers to freed memory to MySQL
All of the options that were wrapped in an option were being freed
before going to MySQL. This is because `Option<CString>::map` takes
ownership of the value, and the raw pointer has no lifetime information.
The simplest solution is to just change the `database.map(|x|
x.as_ptr())` to be `database.as_ref().map(|x| x.as_ptr())`, but I've
gone one step futher and made the API much more difficult to misuse.