Don't assume that integers are always transmitted as network-endian
This is not the long term solution to this problem. I need to
fundamentall rethink how `ToSql` works. I think the biggest indicator
that this is the wrong approach is that we had to stick `type ByteOrder
= NativeEndian` on the `Debug` backend. (Ultimately I'd like to find an
approach that actually shows the output of `std::fmt::Debug::fmt` in
`debug_sql!`).

While I think this will need more work in the future, for now this is
the path of least resistance to continue with the MVP implementation for
MySQL support. MySQL is LittleEndian in its wire protocol, and
NativeEndian in its C API. Since we're just wrapping the C API right
now, I've marked it as `NativeEndian`. Really the fact that we're
wrapping the C API is a detail of the connection, not the backend, but
again I don't think this is a great API long term.

Since SQLite isn't really any endian since it's just calling C
functions, I've moved it to native-endian to do a little bit less work.
Since its bind param handling function is now a bit more unsafe, I've
gone ahead and documented its invariants (these are the requirements for
memory safety, not for the function to be correct)