Implement prepared statements which return data for MySQL
After this change, the core crate's unit test suite is 100% green for
MySQL (`cd diesel && cargo test --features mysql`)

So here's how loading data works. Output values use the same data
structure as input values. For types with a fixed size, we can allocate
the small buffer needed, and everything will just work. For things that
we don't know the size of until we fetch the row, we set the buffer to
size 0.

When mysql hits a field where the buffer isn't large enough, the return
value of `mysql_stmt_fetch` will be `MYSQL_DATA_TRUNCATED`. For each
field that was truncated, the value of `*error` will be set to non-zero
(it's a `my_bool`, so presumably that value is 1. This appears to be the
only error. I do not know why they didn't call it `is_truncated`). It
will also set `*length` to the number of bytes required to hold the
value. At this point we can reallocate the buffer, and try to load that
column directly.

Aside from the general horribleness of this API that I've ranted about
elsewhere, two things in particular greatly bother me here:

- `MYSQL_DATA_TRUNCATED` and all other values mentioned as return values
  for `mysql_stmt_fetch` are `unsigned int`. `mysql_stmt_fetch` returns
  `int`.
- The MySQL client C API is clearly buffering this value on its own. Why
  the fuck can it not just give me that pointer, a byte count, and tell
  me to `memcpy` it before calling `mysql_stmt_fetch` again if I need
  it?!

We can't actually implement `Iterator` for `StatementIterator` because
we want to enforce that the return value of `next` isn't borrowed when
the subsequent call to `next` occurs (sometimes called a "streaming
iterator"). Instead, we implement an almost-iterator API. It'd be nice
to pre-allocate the vec to the appropriate size, but MySQL doesn't give
a way to do that without buffering the entire result set on the client,
which is probably more expensive than what we're doing.

The changes to the `Queryable` macro tests are because `SELECT 1, 2`
returns a big integer in MySQL and there's no way to cast it to a 4 byte
integer.